Change case operation to avoid warning with '-Wsystem-headers' option.
git-svn-id: https://svn.savannah.nongnu.org/svn/avr-libc/trunk/avr-libc@1607 1084037a-409f-4a99-8160-1a4b3577d4c1

+++ include/avr/eeprom.h
@@ -178,9 +178,9 @@ __ATTR_PURE__ static __inline__ uint8_t eeprom_read_byte (const uint8_t *__p)
 {
     do {} while (!eeprom_is_ready ());
 #if	E2END <= 0xFF
-    EEARL = (uint8_t)__p;
+    EEARL = (unsigned)__p;
 #else
-    EEAR = (uint16_t)__p;
+    EEAR = (unsigned)__p;
 #endif
     EECR |= (1 << EERE);
     return EEDR;
@@ -221,9 +221,9 @@ static __inline__ void eeprom_write_byte (uint8_t *__p, uint8_t __value)
     do {} while (!eeprom_is_ready ());
 
 #if	E2END <= 0xFF
-    EEARL = (uint8_t)__p;
+    EEARL = (unsigned)__p;
 #else
-    EEAR = (uint16_t)__p;
+    EEAR = (unsigned)__p;
 #endif
     EEDR = __value;
